Solved the mighty morphin player problem
? In the last build the player would regenerate their appearance in 
every scene because that's what happens when you create a new player. I 
solved this by creating the player in the Preloader and then just adding 
them to each scene when it's created, but this broke the animation 
system for a bit because parts of the player were being destroyed on 
scene change. There's a boolean property for that, ignoreDestroy, which 
fixed that part. This is was a matter of readding input events per scene 
and we seem to have a stable player at last.

? Must remember the guards also do this morphing thing. In fact they do 
in the RELEASE version of TAIP2 which is funny/embarrassing. (Added to 
todos)