Fixed careless whispers, confirmed the queue moves, updated todos
? Just a few tweaks such that the entire game is navigable now it seems. 
(Removing the whisper feature seemed to take care of that.)

? Confirmed the queue is actually a bit more sophisticated than I 
remembered and does move up properly on its own which is less work for 
me. Though how it's implemented may mean it's less easy? Is it tweens of 
velocity? I should really look...

? Okay looking at it it is velocity based but it'll need to be thought 
through a little more than the way it is now which doesn't assume a 
player might be in the way of everyone. Still, there's material.